// src/core/http/generated/models/subscription.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Petvex Admin API
 * OpenAPI spec version: 1.0.0
 */
import type { Plan } from './plan.ts';
import type { SubscriptionBillingCycle } from './subscriptionBillingCycle.ts';
import type { SubscriptionStatus } from './subscriptionStatus.ts';
import type { Tenant } from './tenant.ts';

export interface Subscription {
  id: string;
  tenant_id: string;
  tenant?: Tenant;
  plan_id: string;
  plan?: Plan;
  status: SubscriptionStatus;
  billing_cycle: SubscriptionBillingCycle;
  price: number;
  gateway?: string;
  trial_ends_at?: string;
  current_period_starts_at?: string;
  current_period_ends_at?: string;
  next_billing_at?: string;
  canceled_at?: string;
  suspended_at?: string;
  created_at: string;
  updated_at?: string;
}